# openapi.yaml
openapi: 3.0.1
info:
  title: OSHU API
  description: "유성구 소상공인 지도, 실시간 홍보, 타임세일 API"
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Generated server url
security:
- bearerAuth: []
tags:
- name: Owner Promotions
  description: 점주 홍보 관리
- name: Stores
  description: 소비자 가게·지도 조회
- name: Owner Time Sales
  description: 점주 타임세일 관리
- name: Promotions
  description: 홍보 포스터와 행사 조회
- name: Auth
  description: 회원가입 및 로그인
- name: Owner Stores
  description: 점주 가게 관리
paths:
  /owner/stores:
    get:
      tags:
      - Owner Stores
      summary: 내 가게 목록 조회
      operationId: myStores
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StoreCardResponse"
      security:
      - bearerAuth: []
    post:
      tags:
      - Owner Stores
      summary: 내 가게 등록
      description: openingHours에는 영업 시작·마감 시간을 `09:00 - 21:00` 형식으로 입력합니다.
      operationId: createStore
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StoreCreateRequest"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/StoreDetailResponse"
      security:
      - bearerAuth: []
  /owner/stores/{storeId}/time-sales:
    post:
      tags:
      - Owner Time Sales
      summary: 내 가게 타임세일 등록
      operationId: createTimeSale
      parameters:
      - name: storeId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TimeSaleRequest"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TimeSaleResponse"
      security:
      - bearerAuth: []
  /owner/stores/{storeId}/promotions:
    post:
      tags:
      - Owner Promotions
      summary: 내 가게 홍보 등록
      operationId: createPromotion
      parameters:
      - name: storeId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PromotionRequest"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PromotionResponse"
      security:
      - bearerAuth: []
  /auth/signup:
    post:
      tags:
      - Auth
      summary: 회원가입
      operationId: signUp
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignUpRequest"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/login:
    post:
      tags:
      - Auth
      summary: 로그인
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TokenResponse"
  /owner/time-sales/{timeSaleId}:
    patch:
      tags:
      - Owner Time Sales
      summary: 내 가게 타임세일 수정
      operationId: updateTimeSale
      parameters:
      - name: timeSaleId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TimeSaleRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TimeSaleResponse"
      security:
      - bearerAuth: []
  /owner/time-sales/{timeSaleId}/close:
    patch:
      tags:
      - Owner Time Sales
      summary: 내 가게 타임세일 종료
      operationId: closeTimeSale
      parameters:
      - name: timeSaleId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TimeSaleResponse"
      security:
      - bearerAuth: []
  /owner/stores/{storeId}:
    get:
      tags:
      - Owner Stores
      summary: 내 가게 상세 조회
      operationId: myStore
      parameters:
      - name: storeId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/StoreDetailResponse"
      security:
      - bearerAuth: []
    patch:
      tags:
      - Owner Stores
      summary: 내 가게 정보 수정
      description: openingHours에는 영업 시작·마감 시간을 `09:00 - 21:00` 형식으로 입력합니다.
      operationId: updateStore
      parameters:
      - name: storeId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StoreUpdateRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/StoreDetailResponse"
      security:
      - bearerAuth: []
  /owner/stores/{storeId}/crowd-status:
    patch:
      tags:
      - Owner Stores
      summary: 혼잡도 갱신
      operationId: updateCrowd
      parameters:
      - name: storeId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CrowdStatusRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CrowdStatusResponse"
      security:
      - bearerAuth: []
  /owner/promotions/{promotionId}:
    delete:
      tags:
      - Owner Promotions
      summary: 내 가게 홍보 삭제
      operationId: deletePromotion
      parameters:
      - name: promotionId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: No Content
      security:
      - bearerAuth: []
    patch:
      tags:
      - Owner Promotions
      summary: 내 가게 홍보 수정
      operationId: updatePromotion
      parameters:
      - name: promotionId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PromotionRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PromotionResponse"
      security:
      - bearerAuth: []
  /stores:
    get:
      tags:
      - Stores
      summary: 가게 목록 조회
      operationId: list
      parameters:
      - name: keyword
        in: query
        required: false
        schema:
          type: string
      - name: category
        in: query
        required: false
        schema:
          type: string
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 20
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageResponseStoreCardResponse"
  /stores/{storeId}:
    get:
      tags:
      - Stores
      summary: 가게 상세 조회
      operationId: detail
      parameters:
      - name: storeId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/StoreDetailResponse"
  /stores/{storeId}/summary:
    get:
      tags:
      - Stores
      summary: 지도 가게 선택 요약
      operationId: summary
      parameters:
      - name: storeId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/StoreCardResponse"
  /stores/{storeId}/promotions:
    get:
      tags:
      - Stores
      summary: 가게 행사 조회
      operationId: promotions
      parameters:
      - name: storeId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PromotionResponse"
  /stores/{storeId}/crowd-status:
    get:
      tags:
      - Stores
      summary: 가게 혼잡도 조회
      operationId: crowdStatus
      parameters:
      - name: storeId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CrowdStatusResponse"
  /stores/map:
    get:
      tags:
      - Stores
      summary: 지도 메인
      description: 공공데이터 키가 설정되면 반경 내 공공 상가 정보를 우선 반환합니다.
      operationId: map
      parameters:
      - name: latitude
        in: query
        required: true
        schema:
          type: number
          format: double
      - name: longitude
        in: query
        required: true
        schema:
          type: number
          format: double
      - name: radius
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 1500
      - name: timeSaleOnly
        in: query
        required: false
        schema:
          type: boolean
          default: false
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StoreCardResponse"
  /promotions:
    get:
      tags:
      - Promotions
      summary: 홍보 포스터 목록 조회
      operationId: list_1
      parameters:
      - name: status
        in: query
        required: false
        schema:
          type: string
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 20
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageResponsePromotionResponse"
  /promotions/{promotionId}:
    get:
      tags:
      - Promotions
      summary: 행사 상세 조회
      operationId: detail_1
      parameters:
      - name: promotionId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PromotionResponse"
components:
  schemas:
    StoreCreateRequest:
      required:
      - address
      - category
      - latitude
      - longitude
      - name
      type: object
      properties:
        name:
          type: string
          description: 가게 이름
          example: 오슈 베이커리
        category:
          type: string
          description: 가게 업종
          example: 카페
        description:
          type: string
          description: 가게 소개
          example: 갓 구운 빵과 커피를 판매합니다.
        address:
          type: string
          description: 가게 주소
          example: 대전광역시 유성구 궁동 123
        latitude:
          type: number
          description: 위도
          format: double
          example: 36.3622
        longitude:
          type: number
          description: 경도
          format: double
          example: 127.3449
        phone:
          type: string
          description: 가게 전화번호
          example: 042-000-0001
        openingHours:
          type: string
          description: 가게 영업시간. 시작 시간과 마감 시간을 함께 입력합니다.
          example: 09:00 - 21:00
      description: 가게 등록 요청
    CrowdStatusResponse:
      type: object
      properties:
        level:
          type: string
        label:
          type: string
        estimatedWaitingMinutes:
          type: integer
          format: int32
    PromotionResponse:
      type: object
      properties:
        promotionId:
          type: integer
          format: int64
        storeId:
          type: integer
          format: int64
        storeName:
          type: string
        type:
          type: string
        title:
          type: string
        content:
          type: string
        imageUrl:
          type: string
        startAt:
          type: string
          format: date-time
        endAt:
          type: string
          format: date-time
        status:
          type: string
    StoreDetailResponse:
      type: object
      properties:
        storeId:
          type: integer
          format: int64
        name:
          type: string
        category:
          type: string
        description:
          type: string
        address:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        phone:
          type: string
        openingHours:
          type: string
        crowdStatus:
          $ref: "#/components/schemas/CrowdStatusResponse"
        promotions:
          type: array
          items:
            $ref: "#/components/schemas/PromotionResponse"
        timeSales:
          type: array
          items:
            $ref: "#/components/schemas/TimeSaleResponse"
    TimeSaleResponse:
      type: object
      properties:
        timeSaleId:
          type: integer
          format: int64
        storeId:
          type: integer
          format: int64
        productName:
          type: string
        originalPrice:
          type: integer
          format: int32
        salePrice:
          type: integer
          format: int32
        startAt:
          type: string
          format: date-time
        endAt:
          type: string
          format: date-time
        notice:
          type: string
        status:
          type: string
    TimeSaleRequest:
      required:
      - endAt
      - originalPrice
      - productName
      - salePrice
      - startAt
      type: object
      properties:
        productName:
          type: string
        originalPrice:
          type: integer
          format: int32
        salePrice:
          type: integer
          format: int32
        startAt:
          type: string
          format: date-time
        endAt:
          type: string
          format: date-time
        notice:
          type: string
    PromotionRequest:
      required:
      - endAt
      - startAt
      - title
      - type
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        content:
          type: string
        imageUrl:
          type: string
        startAt:
          type: string
          format: date-time
        endAt:
          type: string
          format: date-time
    SignUpRequest:
      required:
      - loginId
      - password
      type: object
      properties:
        loginId:
          type: string
        password:
          type: string
    MessageResponse:
      type: object
      properties:
        message:
          type: string
    LoginRequest:
      required:
      - loginId
      - password
      type: object
      properties:
        loginId:
          type: string
        password:
          type: string
    TokenResponse:
      type: object
      properties:
        accessToken:
          type: string
        tokenType:
          type: string
    StoreUpdateRequest:
      type: object
      properties:
        description:
          type: string
          description: 가게 소개
          example: 갓 구운 빵과 커피를 판매합니다.
        phone:
          type: string
          description: 가게 전화번호
          example: 042-000-0001
        openingHours:
          type: string
          description: 가게 영업시간. 시작 시간과 마감 시간을 함께 입력합니다.
          example: 09:00 - 21:00
      description: 가게 정보 수정 요청
    CrowdStatusRequest:
      required:
      - estimatedWaitingMinutes
      - level
      type: object
      properties:
        level:
          type: string
          enum:
          - RELAXED
          - NORMAL
          - BUSY
          - VERY_BUSY
        estimatedWaitingMinutes:
          type: integer
          format: int32
    PageResponseStoreCardResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/StoreCardResponse"
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
    StoreCardResponse:
      type: object
      properties:
        storeId:
          type: integer
          format: int64
        name:
          type: string
        category:
          type: string
        address:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        crowdLevel:
          type: string
        timeSaleActive:
          type: boolean
        externalData:
          type: boolean
    PageResponsePromotionResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/PromotionResponse"
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT